Refactored model class view into separate
