--- app/forms.py
@@ -1,5 +1,0 @@
-from django import forms
-
-
-class DocumentForm(forms.Form):
-    file = forms.FileField(label='select file to upload')
